refactor!: move autonat into separate package (#2107)
