+++ packages/libp2p/src/errors.ts
@@ -37,7 +37,6 @@ export enum codes {
   ERR_INVALID_PEER = 'ERR_INVALID_PEER',
   ERR_MUXER_UNAVAILABLE = 'ERR_MUXER_UNAVAILABLE',
   ERR_NOT_FOUND = 'ERR_NOT_FOUND',
-  ERR_TIMEOUT = 'ERR_TIMEOUT',
   ERR_TRANSPORT_UNAVAILABLE = 'ERR_TRANSPORT_UNAVAILABLE',
   ERR_TRANSPORT_DIAL_FAILED = 'ERR_TRANSPORT_DIAL_FAILED',
   ERR_UNSUPPORTED_PROTOCOL = 'ERR_UNSUPPORTED_PROTOCOL',
