+++ packages/protocol-autonat/package.json
@@ -1,0 +1,71 @@
+{
+  "name": "@libp2p/autonat",
+  "version": "1.0.0",
+  "description": "Implementation of Autonat Protocol",
+  "license": "Apache-2.0 OR MIT",
+  "homepage": "https://github.com/libp2p/js-libp2p/tree/master/packages/protocol-autonat#readme",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/libp2p/js-libp2p.git"
+  },
+  "bugs": {
+    "url": "https://github.com/libp2p/js-libp2p/issues"
+  },
+  "type": "module",
+  "types": "./dist/src/index.d.ts",
+  "files": [
+    "src",
+    "dist",
+    "!dist/test",
+    "!**/*.tsbuildinfo"
+  ],
+  "exports": {
+    ".": {
+      "types": "./dist/src/index.d.ts",
+      "import": "./dist/src/index.js"
+    }
+  },
+  "eslintConfig": {
+    "extends": "ipfs",
+    "parserOptions": {
+      "sourceType": "module"
+    }
+  },
+  "scripts": {
+    "start": "node dist/src/main.js",
+    "build": "aegir build",
+    "test": "aegir test",
+    "clean": "aegir clean",
+    "generate": "protons ./src/pb/index.proto",
+    "lint": "aegir lint",
+    "test:chrome": "aegir test -t browser --cov",
+    "test:chrome-webworker": "aegir test -t webworker",
+    "test:firefox": "aegir test -t browser -- --browser firefox",
+    "test:firefox-webworker": "aegir test -t webworker -- --browser firefox",
+    "test:node": "aegir test -t node --cov",
+    "dep-check": "aegir dep-check"
+  },
+  "dependencies": {
+    "@libp2p/interface": "^0.1.2",
+    "@libp2p/interface-internal": "^0.1.5",
+    "@libp2p/peer-id": "^3.0.2" ,
+    "@libp2p/peer-id-factory": "^3.0.4" ,
+    "@libp2p/logger": "^3.0.2",
+    "@multiformats/multiaddr": "^12.1.5",
+    "it-first": "^3.0.1",
+    "it-length-prefixed": "^9.0.1",
+    "it-map": "^3.0.3",
+    "it-parallel": "^3.0.0",
+    "it-pipe": "^3.0.1",
+    "private-ip": "^3.0.0",
+    "protons-runtime": "^5.0.0",
+    "uint8arraylist": "^2.4.3"
+  },
+  "devDependencies": {
+    "aegir": "^41.0.2",
+    "it-all": "^3.0.1",
+    "it-pushable": "^3.2.0",
+    "sinon": "^17.0.0",
+    "sinon-ts": "^1.0.0"
+  }
+}
