+++ packages/protocol-autonat/typedoc.json
@@ -1,0 +1,5 @@
+{
+  "entryPoints": [
+    "./src/index.ts"
+  ]
+}
